+++ google_play_scraper/__init__.py
@@ -2,4 +2,4 @@ from .constants.google_play import Sort
 from .features.app import app
 from .features.reviews import reviews, reviews_all
 
-VERSION = __version__ = "0.0.3.1"
+VERSION = __version__ = "0.0.3.2"